Fix: Information related to tokens in README
Fix documentation related to tokens. The `secrets.GITHUB_TOKEN` provided by the GitHub Actions App can do everything related to the repo if we elevate it's permissions, including calling `workflow_dispatch` and `repository_dispatch`.

Some people in my organization are using PAT's instead of the `secrets.GITHUB_TOKEN` when using this action because of this README suggestion, even when they are calling the same repository.